Move remaining object modules into `objects`
This leaves the `geometry` module a bit light, but I don't know where to
put its remaining sub-module. Over time, a solution will present itself,
I'm sure.
