+++ crates/fj-kernel/src/algorithms/intersection/surface_surface.rs
@@ -1,6 +1,6 @@
 use fj_math::{Line, Point, Scalar, Vector};
 
-use crate::geometry::{Curve, Surface};
+use crate::objects::{Curve, Surface};
 
 /// Test intersection between two surfaces
 pub fn surface_surface(a: &Surface, b: &Surface) -> Option<Curve<3>> {
@@ -60,7 +60,7 @@ fn extract_plane(surface: &Surface) -> (Vector<3>, Scalar) {
 mod tests {
     use fj_math::Transform;
 
-    use crate::geometry::{Curve, Surface};
+    use crate::objects::{Curve, Surface};
 
     use super::surface_surface;
