+++ crates/fj-kernel/src/objects/mod.rs
@@ -39,14 +39,18 @@
 //! [`Handle`]: crate::shape::Handle
 //! [`Shape`]: crate::shape::Shape
 
+mod curves;
 mod cycle;
 mod edge;
 mod face;
+mod surfaces;
 mod vertex;
 
 pub use self::{
+    curves::Curve,
     cycle::Cycle,
     edge::{Edge, VerticesOfEdge},
     face::{CyclesInFace, Face},
+    surfaces::{Surface, SweptCurve},
     vertex::Vertex,
 };
