+++ crates/fj-kernel/src/shape/mapping.rs
@@ -1,9 +1,6 @@
 use std::collections::HashMap;
 
-use crate::{
-    geometry::{Curve, Surface},
-    objects::{Cycle, Edge, Face, Vertex},
-};
+use crate::objects::{Curve, Cycle, Edge, Face, Surface, Vertex};
 
 use super::Handle;
